requirements.txt
httpx>=0.24.0
click>=8.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
